Extract rtcombat into own module.
